Move logic printCoverage to common
