+++ .github/workflows/coverage-badge.yml
@@ -3,8 +3,6 @@ name: Create coverage badge
 on:
   push:
     branches: [ main ]
-  pull_request:
-    branches: [ main ]
 
 env:
   JAVA_VERSION: '17'
